Projects/Small/(W.I.P) Self-HostedSimpleDiscordBot (.py)/W.I.P/Assets/requirements.txt
discord.py==2.3.0
getmac==0.9.4
pycryptodome==3.18.0
py-cpuinfo==9.0.0